radeon: Sane handling of timeouts in WaitForVerticalSync(2).
RADEONWaitForVerticalSync() and RADEONWaitForVerticalSync2() need to wait
for a timeout specified in milliseconds; looping around usleep() causes
the timeout to be unnecessarily long, as the OS may sleep longer than
requested (on Linux the minimum actual sleep value may be several ms).

The new logic uses gettimeofday() in the loop to see when the (absolute)
timeout has arrived.

Signed-off-by: Alon Ziv <alonz@nolaviz.org>